+++ examples/examples/surf-client.rs
@@ -54,10 +54,13 @@ async fn run_query() -> cynic::GraphQLResponse<FilmDirectorQuery> {
 }
 
 fn build_query() -> cynic::Operation<'static, FilmDirectorQuery> {
-    use cynic::QueryFragment;
-    cynic::Operation::query(FilmDirectorQuery::fragment(&FilmArguments {
-        id: Some("ZmlsbXM6MQ==".into()),
-    }))
+    use cynic::{FragmentContext, QueryFragment};
+
+    cynic::Operation::query(FilmDirectorQuery::fragment(FragmentContext::with_args(
+        &FilmArguments {
+            id: Some("ZmlsbXM6MQ==".into()),
+        },
+    )))
 }
 
 #[cfg(test)]